+++ src/main/java/com/andy/recipe/like/domain/Like.java
@@ -3,24 +3,24 @@ package com.andy.recipe.like.domain;
 import java.sql.Timestamp;
 
 public class Like {
-	private int id;
-	private int userId;
+	private long id;
+	private long userId;
 	private int postId;
 	private Timestamp createdAt;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
 
@@ -39,4 +39,5 @@ public class Like {
 	public void setCreatedAt(Timestamp createdAt) {
 		this.createdAt = createdAt;
 	}
+
 }
